fm-lifecycle-orchestrator/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
python-dotenv==1.0.1
openai==1.51.2
langchain==0.3.3
langchain-openai==0.2.2
pypdf2==3.0.1
pdfplumber==0.11.4
pytesseract==0.3.13
pillow==10.4.0
python-docx==1.1.2
httpx==0.27.2
